Construct PCAGenomicSignatures object

The default contents of PCAGenomicSignatures object, with a set of accessor and setter generic functions. When you create this object, colData(x)$studies should be populated before adding any information in trainingData(x) slot

assay(x) : PCAmodel (= avgLoading) containing genes x PCclusters

metadata(x)$cluster : A vector of integers (from 1:k) indicating the cluster to which each point is allocated. Check stats::kmeans function.

metadata(x)$size : The number of points in each cluster. Check stats::kmeans function.

metadata(x)$iter : The number of (outer) iterations. Check stats::kmeans function.

metadata(x)$k : The number of PCclusters.

metadata(x)$n : The number of top PCs from each dataset.

metadata(x)$geneSets : Name of the prior gene sets used to annotate average loadings.

colData(x)$studies : A list of character vectors containing studies contributing to each PC cluster.

colData(x)$gsea : A list of gseaResult objects. Build using clusterProfiler::GSEA function.

PCAGenomicSignatures(...)

Arguments

...
Additional arguments for supporting functions