Construct PCAGenomicSignatures object

The default contents of PCAGenomicSignatures object, with a set of accessor and setter generic functions, which extract either the assay, colData, metadata, or trainingData slots of a PCAGenomicSignatures-class object. When you create this object, colData$studies should be populated before adding any information in trainingData slot

PCAGenomicSignatures(...)

Value

PCAGenomicSignatures object with multiple setters or accessors

Details

model(x) : PCAmodel (= avgLoading) containing genes x PCclusters

metadata(x)$cluster : A vector of integers (from 1:k) indicating the cluster to which each point is allocated. Check stats::kmeans function.

metadata(x)$size : The number of points in each cluster. Check stats::kmeans function.

metadata(x)$iter : The number of (outer) iterations. Check stats::kmeans function.

metadata(x)$k : The number of PCclusters.

metadata(x)$n : The number of top PCs from each dataset.

metadata(x)$geneSets : Name of the prior gene sets used to annotate average loadings.

colData(x)$studies : A list of character vectors containing studies contributing to each PC cluster.

colData(x)$gsea : A list of gseaResult objects. Build using clusterProfiler::GSEA function.

Slots

trainingData

A DataFrame class object for metadata associated with training data

Setters

Setter method values (i.e., function(x) <- value):

geneSets<- : A character vector containing the name of gene sets used to annotate average loadings

studies<- : A list of character vectors containing gene sets used to annotate average loadings

gsea<- : A list of gseaResult objects.

metadata<- : A list object of metadata

`$<-` : A vector to replace the indicated column in colData

Accessors

All the accessors inherited from SummarizedExperiment are available and the additional accessors for PCAGenomicSignatures specific data are listed below.

model : Equivalent to the assays(x)$model accessor for convenience

geneSets : Access the metadata(x)$geneSets slot

studies : Access the colData(x)$studies slot

gsea : Access the colData(x)$gsea

`$` : Access a column in colData

trainingData : Access the trainingData slot

mesh : Access the trainingData(x)$MeSH slot

PCAsummary : Access the trainingData(x)$PCAsummary slot